(PC-14559) feat(SuspendedAccount): Refactor components in appropriate folders
