+++ src/features/auth/suspendedAccount/FraudulentAccount/tests/FraudulentAccount.web.test.tsx
